chore(apollo-vertex): add hooks to registry
