+++ apps/apollo-vertex/package.json
@@ -55,6 +55,7 @@
     "@tailwindcss/postcss": "^4.1.17",
     "@tanstack/react-query": "^5.90.21",
     "@tanstack/react-table": "^8.21.3",
+    "@uidotdev/usehooks": "^2.4.1",
     "@vercel/analytics": "^1.5.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
@@ -86,26 +87,6 @@
     "tailwindcss": "^4.1.17",
     "vaul": "^1.1.2",
     "zod": "^4.3.5"
-  },
-  "peerDependencies": {
-    "@tanstack/react-db": ">=0",
-    "@uipath/data-formatting": ">=0",
-    "@uipath/vs-core": ">=0",
-    "luxon": ">=3"
-  },
-  "peerDependenciesMeta": {
-    "@tanstack/react-db": {
-      "optional": true
-    },
-    "@uipath/data-formatting": {
-      "optional": true
-    },
-    "@uipath/vs-core": {
-      "optional": true
-    },
-    "luxon": {
-      "optional": true
-    }
   },
   "devDependencies": {
     "@biomejs/biome": "^2.3.6",
